+++ common/runtime/init.go
@@ -8,6 +8,7 @@ import (
 	"github.com/turt2live/matrix-media-repo/common/rcontext"
 	"github.com/turt2live/matrix-media-repo/common/version"
 	"github.com/turt2live/matrix-media-repo/ipfs_proxy"
+	"github.com/turt2live/matrix-media-repo/plugins"
 	"github.com/turt2live/matrix-media-repo/storage"
 	"github.com/turt2live/matrix-media-repo/storage/datastore"
 	"github.com/turt2live/matrix-media-repo/storage/datastore/ds_s3"
@@ -19,6 +20,7 @@ func RunStartupSequence() {
 	config.CheckDeprecations()
 	LoadDatabase()
 	LoadDatastores()
+	plugins.ReloadPlugins()
 
 	logrus.Info("Starting IPFS (if enabled)...")
 	ipfs_proxy.Reload()