+++ matrix/client_server.go
@@ -72,17 +72,3 @@ func doRequest(ctx rcontext.RequestContext, method string, urlStr string, body i
 
 	return nil
 }
-
-func makeUrl(parts ...string) string {
-	res := ""
-	for i, p := range parts {
-		if p[len(p)-1:] == "/" {
-			res += p[:len(p)-1]
-		} else if p[0] != '/' && i > 0 {
-			res += "/" + p
-		} else {
-			res += p
-		}
-	}
-	return res
-}