+++ CHANGELOG.md
@@ -7,7 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-*Nothing yet.*
+### Added
+
+* Introduced early plugin support (only for antispam for now).
+  * Includes a simple OCR plugin to help mitigate text-based image spam.
 
 ## [1.2.2] - December 8th, 2020